refactor(runner-core): split path projection facade
